menambahkan gitignore

+++ models/field.go
@@ -20,7 +20,7 @@ type Field struct {
 	Location    string      `gorm:"column:location;not null" json:"location"`
 	Description string      `gorm:"column:description" json:"description"`
 	Type        FieldType   `gorm:"column:type;not null" json:"type"`
-	Price       int         `gorm:"column:price;not null" json:"price"`
+	Price       float64     `gorm:"column:price;not null" json:"price"`
 	Facilities  string      `gorm:"column:facilities" json:"facilities"`
 	CreatedAt   time.Time   `gorm:"column:created_at;autoCreateTime" json:"created_at"`
 	UpdatedAt   time.Time   `gorm:"column:updated_at;autoUpdateTime" json:"updated_at"`
